Migrated goodFeaturesToTrack to features module.
