+++ modules/imgproc/misc/java/gen_dict.json
@@ -105,7 +105,6 @@
         }
     },
     "func_arg_fix" : {
-        "goodFeaturesToTrack" : { "corners" : {"ctype" : "vector_Point"} },
         "minEnclosingCircle" : { "points" : {"ctype" : "vector_Point2f"} },
         "fitEllipse" : { "points" : {"ctype" : "vector_Point2f"} },
         "fillPoly"   : { "pts" : {"ctype" : "vector_vector_Point"} },
