+++ modules/imgproc/misc/java/test/ImgprocTest.java
@@ -1007,26 +1007,6 @@ public class ImgprocTest extends OpenCVTestCase {
         assertMatEqual(truth, dst);
     }
 
-    public void testGoodFeaturesToTrackMatListOfPointIntDoubleDouble() {
-        Mat src = gray0;
-        Imgproc.rectangle(src, new Point(2, 2), new Point(8, 8), new Scalar(100), -1);
-        MatOfPoint lp = new MatOfPoint();
-
-        Imgproc.goodFeaturesToTrack(src, lp, 100, 0.01, 3);
-
-        assertEquals(4, lp.total());
-    }
-
-    public void testGoodFeaturesToTrackMatListOfPointIntDoubleDoubleMatIntBooleanDouble() {
-        Mat src = gray0;
-        Imgproc.rectangle(src, new Point(2, 2), new Point(8, 8), new Scalar(100), -1);
-        MatOfPoint lp = new MatOfPoint();
-
-        Imgproc.goodFeaturesToTrack(src, lp, 100, 0.01, 3, gray1, 4, 3, true, 0);
-
-        assertEquals(4, lp.total());
-    }
-
     public void testGrabCutMatMatRectMatMatInt() {
         fail("Not yet implemented");
     }
