+++ modules/imgproc/misc/js/gen_dict.json
@@ -59,7 +59,6 @@
             "getRectSubPix",
             "getRotationMatrix2D",
             "getStructuringElement",
-            "goodFeaturesToTrack",
             "grabCut",
             "HoughCircles",
             "HoughLines",
