+++ modules/imgproc/misc/objc/gen_dict.json
@@ -45,7 +45,6 @@
     },
     "func_arg_fix" : {
         "Imgproc" : {
-            "goodFeaturesToTrack" : { "corners" : {"ctype" : "vector_Point"} },
             "minEnclosingCircle" : { "points" : {"ctype" : "vector_Point2f"} },
             "fitEllipse" : { "points" : {"ctype" : "vector_Point2f"} },
             "fillPoly"   : { "pts" : {"ctype" : "vector_vector_Point"},
