+++ samples/cpp/lkdemo.cpp
@@ -1,5 +1,6 @@
 #include "opencv2/video/tracking.hpp"
 #include "opencv2/imgproc.hpp"
+#include "opencv2/features.hpp"
 #include "opencv2/videoio.hpp"
 #include "opencv2/highgui.hpp"
