+++ samples/cpp/tutorial_code/TrackingMotion/cornerSubPix_Demo.cpp
@@ -6,6 +6,7 @@
 
 #include "opencv2/highgui.hpp"
 #include "opencv2/imgproc.hpp"
+#include "opencv2/features.hpp"
 #include <iostream>
 
 using namespace cv;
