+++ samples/cpp/tutorial_code/video/optical_flow/optical_flow.cpp
@@ -2,6 +2,7 @@
 #include <opencv2/core.hpp>
 #include <opencv2/highgui.hpp>
 #include <opencv2/imgproc.hpp>
+#include <opencv2/features.hpp>
 #include <opencv2/videoio.hpp>
 #include <opencv2/video.hpp>
