+++ samples/gpu/pyrlk_optical_flow.cpp
@@ -9,6 +9,7 @@
 #include <opencv2/cudaoptflow.hpp>
 #include <opencv2/cudaimgproc.hpp>
 #include <opencv2/cudaarithm.hpp>
+#include <opencv2/cudafeatures2d.hpp>
 
 using namespace std;
 using namespace cv;
