+++ samples/java/tutorial_code/TrackingMotion/good_features_to_track/GoodFeaturesToTrackDemo.java
@@ -20,6 +20,7 @@ import org.opencv.core.Scalar;
 import org.opencv.highgui.HighGui;
 import org.opencv.imgcodecs.Imgcodecs;
 import org.opencv.imgproc.Imgproc;
+import org.opencv.features.Features;
 
 class GoodFeaturesToTrack {
     private Mat src = new Mat();
@@ -99,7 +100,7 @@ class GoodFeaturesToTrack {
         Mat copy = src.clone();
 
         /// Apply corner detection
-        Imgproc.goodFeaturesToTrack(srcGray, corners, maxCorners, qualityLevel, minDistance, new Mat(),
+        Features.goodFeaturesToTrack(srcGray, corners, maxCorners, qualityLevel, minDistance, new Mat(),
                 blockSize, gradientSize, useHarrisDetector, k);
 
         /// Draw corners detected
