+++ samples/tapi/pyrlk_optical_flow.cpp
@@ -4,6 +4,7 @@
 
 #include "opencv2/core/utility.hpp"
 #include "opencv2/imgcodecs.hpp"
+#include "opencv2/features.hpp"
 #include "opencv2/videoio.hpp"
 #include "opencv2/highgui.hpp"
 #include "opencv2/core/ocl.hpp"
